Separate into two controller classes
