+++ gigadb/app/tools/dropbox-generator/commands/WasabiBucketController.php
@@ -12,14 +12,12 @@ use yii\console\Controller;
 use yii\console\ExitCode;
 use Aws\S3\Exception\S3Exception;
 use Aws\S3\S3Client;
-use Aws\Iam\Exception\IamException;
-use Aws\Iam\IamClient;
 
 /**
  * Read file in bucket
  *
  */
-class WasabiController extends Controller
+class WasabiBucketController extends Controller
 {
     /**
      * For storing credentials to access Wasabi
@@ -43,13 +41,6 @@ class WasabiController extends Controller
     public $filePath = '';
 
     /**
-     * Manuscript identifier
-     *
-     * @var string $manuscriptId
-     */
-    public $manuscriptId = '';
-
-    /**
      * Specify options available to console command provided by this controller.
      *
      * @var    $actionID
@@ -60,7 +51,6 @@ class WasabiController extends Controller
         return [
             'bucket',
             'filePath',
-            'manuscriptId',
         ];
     }
 
@@ -78,51 +68,6 @@ class WasabiController extends Controller
             'use_path_style_endpoint' => true,
             // 'debug'   => true
         );
-    }
-
-    /**
-     * Create user account
-     * @return int Exit code
-     */
-    public function actionCreategigadbuser()
-    {
-        // Use manuscript identifier as the username for new user account
-        // Bucket name will therefore be "bucket-$optUserName"
-        $optUserName   = $this->manuscriptId;
-
-        // Return usage unless mandatory options are passed.
-        if ($optUserName === '') {
-            $this->stdout(
-                "\nUsage:\n\t./yii wasabi/creategigadbuser --manuscriptId theManuscriptId" . PHP_EOL
-            );
-            return ExitCode::USAGE;
-        }
-
-        $credentials = array(
-            'credentials' => [
-                'key' => Yii::$app->params['wasabi']['key'],
-                'secret' => Yii::$app->params['wasabi']['secret']
-            ],
-            'endpoint' => Yii::$app->params['wasabi']['iam_endpoint'],
-            'region' => Yii::$app->params['wasabi']['iam_region'],
-            'version' => 'latest',
-            'use_path_style_endpoint' => true,
-        );
-
-        //Establish connection to wasabi
-        $iam = new IamClient($credentials);
-
-        try {
-            // An Aws Result object is returned
-            $result = $iam->createUser([
-                'UserName' => "$optUserName"
-            ]);
-            var_dump($result);
-        } catch (IamException $e) {
-            echo $e->getMessage() . PHP_EOL;
-        }
-
-        return ExitCode::OK;
     }
 
     /**
